pyproject.toml
[tool.poetry]
name = "cipher"
version = "0.1.0"
description = "A backtesting framework."
authors = ["Oleksandr Polieno <polyenoom@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.4"
typer = "^0.7.0"
pandas-ta = "^0.3.14b0"
ujson = "^5.6.0"
requests = "^2.28.1"
jupyterlab = "^3.5.2"
mplfinance = "^0.12.9b7"

[tool.poetry.group.dev.dependencies]
black = {extras = ["d"], version = "^22.12.0"}
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
cipher = "cipher.cli:app"
